refactor(cli): move onboarding openclaw and provider config
