+++ src/lib/onboard-remote-provider-config.ts
@@ -1,0 +1,76 @@
+// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import { DEFAULT_CLOUD_MODEL } from "./inference-config";
+
+export const BUILD_ENDPOINT_URL = "https://integrate.api.nvidia.com/v1";
+export const OPENAI_ENDPOINT_URL = "https://api.openai.com/v1";
+export const ANTHROPIC_ENDPOINT_URL = "https://api.anthropic.com";
+export const GEMINI_ENDPOINT_URL = "https://generativelanguage.googleapis.com/v1beta/openai/";
+
+export const REMOTE_PROVIDER_CONFIG = {
+  build: {
+    label: "NVIDIA Endpoints",
+    providerName: "nvidia-prod",
+    providerType: "nvidia",
+    credentialEnv: "NVIDIA_API_KEY",
+    endpointUrl: BUILD_ENDPOINT_URL,
+    helpUrl: "https://build.nvidia.com/settings/api-keys",
+    modelMode: "catalog",
+    defaultModel: DEFAULT_CLOUD_MODEL,
+    skipVerify: true,
+  },
+  openai: {
+    label: "OpenAI",
+    providerName: "openai-api",
+    providerType: "openai",
+    credentialEnv: "OPENAI_API_KEY",
+    endpointUrl: OPENAI_ENDPOINT_URL,
+    helpUrl: "https://platform.openai.com/api-keys",
+    modelMode: "curated",
+    defaultModel: "gpt-5.4",
+    skipVerify: true,
+  },
+  anthropic: {
+    label: "Anthropic",
+    providerName: "anthropic-prod",
+    providerType: "anthropic",
+    credentialEnv: "ANTHROPIC_API_KEY",
+    endpointUrl: ANTHROPIC_ENDPOINT_URL,
+    helpUrl: "https://console.anthropic.com/settings/keys",
+    modelMode: "curated",
+    defaultModel: "claude-sonnet-4-6",
+  },
+  anthropicCompatible: {
+    label: "Other Anthropic-compatible endpoint",
+    providerName: "compatible-anthropic-endpoint",
+    providerType: "anthropic",
+    credentialEnv: "COMPATIBLE_ANTHROPIC_API_KEY",
+    endpointUrl: "",
+    helpUrl: null,
+    modelMode: "input",
+    defaultModel: "",
+  },
+  gemini: {
+    label: "Google Gemini",
+    providerName: "gemini-api",
+    providerType: "openai",
+    credentialEnv: "GEMINI_API_KEY",
+    endpointUrl: GEMINI_ENDPOINT_URL,
+    helpUrl: "https://aistudio.google.com/app/apikey",
+    modelMode: "curated",
+    defaultModel: "gemini-2.5-flash",
+    skipVerify: true,
+  },
+  custom: {
+    label: "Other OpenAI-compatible endpoint",
+    providerName: "compatible-endpoint",
+    providerType: "openai",
+    credentialEnv: "COMPATIBLE_API_KEY",
+    endpointUrl: "",
+    helpUrl: null,
+    modelMode: "input",
+    defaultModel: "",
+    skipVerify: true,
+  },
+};
